+++ src/hooks/useVimeoPlayer.ts
@@ -2,7 +2,7 @@ import { useEffect, useMemo, useRef } from 'react';
 import type WebView from 'react-native-webview';
 import VimeoPlayerInstance from '../module/VimeoPlayerInstance';
 import type { VimeoSource } from '../types';
-import type { VimeoPlayerOptions } from '../types/iframe';
+import type { VimeoPlayerOptions } from '../types/vimeo';
 
 /**
  * @param source - The source of the Vimeo video. It can be a string or an object with a `url` property.
