+++ src/styles.css
@@ -1,0 +1,4 @@
+iframe {
+  width: 100%;
+  height: 100%;
+}
